+++ packages/test/src/HTTP.test.ts
@@ -20,7 +20,7 @@ describe("HTTP", () => {
       },
       expected: {
         foo: "bar",
-        "user-agent": `test/9000, cerbos-sdk-javascript-http/${version}`,
+        "user-agent": `test/9000 cerbos-sdk-javascript-http/${version}`,
       },
     },
     {
@@ -31,7 +31,7 @@ describe("HTTP", () => {
       }),
       expected: {
         foo: "bar",
-        "user-agent": `test/9000, cerbos-sdk-javascript-http/${version}`,
+        "user-agent": `test/9000 cerbos-sdk-javascript-http/${version}`,
       },
     },
     {